+++ Infrastructure/Services/CarService.cs
@@ -1,3 +1,6 @@
+using System.Net;
+using Infrastructure.ApiResponse;
+
 namespace Infrastructure.Services;
 
 using System.Collections.Generic;
@@ -14,44 +17,62 @@ public class CarService : ICarService
     _context = context;
   }
 
-  public  List<Car> GetAll()
+  public  Response<List<Car>> GetAll()
   {
     using var context = _context.Connection();
     string cmd = "SELECT * FROM Cars";
     var cars = context.Query<Car>(cmd).ToList();
-    return cars;
+    return new Response<List<Car>>(cars);
   }
 
-  public Car GetCarById(int id)
+  public Response<Car> GetCarById(int id)
   {
     using var context = _context.Connection();
     string cmd = "SELECT * FROM Cars WHERE id = @id";
     var cars = context.QueryFirstOrDefault<Car>(cmd,new {Id = id});
-    return cars;
+    return cars == null
+      ? new Response<Car>(HttpStatusCode.NotFound, "Cars Not Found")
+      : new Response<Car>(cars);
   }
 
 
-  public bool CreateCar(Car car)
+  public Response<bool> CreateCar(Car car)
   {
     using var context = _context.Connection();
     var cmd = "INSERT INTO Cars (Model, Manufacturer, Year, PricePerDay) VALUES (@Model, @Manufacturer, @Year, @PricePerDay)";
     var response = context.Execute(cmd, car);
-    return response > 0;
+    return response == 0
+      ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal Server Error")
+      : new Response<bool>(HttpStatusCode.Created, "Car successfully created");
   }
 
-  public bool UpdateCar(Car car)
+  public Response<bool> UpdateCar(Car car)
   {
     using var context = _context.Connection();
+    var existingCar = GetCarById(car.CarId).Data;
+    if (existingCar == null)
+    {
+      return new Response<bool>(HttpStatusCode.NotFound, "Car not found");
+    }
     var cmd = "UPDATE Cars SET Model = @Model, Manufacturer = @Manufacturer, Year = @Year, PricePerDay = @PricePerDay WHERE CarId = @CarId";
     var response = context.Execute(cmd, car);
-    return response > 0;
+    return response > 0
+      ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal server error")
+      : new Response<bool>(HttpStatusCode.OK, "Car successfully updated");
   }
 
-  public bool DeleteCar(int id)
+  public Response<bool> DeleteCar(int id)
   {
     using var context = _context.Connection();
+    var car = GetCarById(id).Data;
+    if (car == null)
+    {
+      return new Response<bool>(HttpStatusCode.NotFound, "Car not found");
+    }
     var cmd = "DELETE FROM Cars WHERE CarId = @CarId";
     var response = context.Execute(cmd, new { CarId = id });
-    return response != 0;
+    return response > 0
+      ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal server error")
+      : new Response<bool>(HttpStatusCode.OK, "Car successfully deleted");
   }
 }
