+++ Infrastructure/Services/CustomerService.cs
@@ -1,3 +1,6 @@
+using System.Net;
+using Infrastructure.ApiResponse;
+
 namespace Infrastructure.Services;
 
 using System.Collections.Generic;
@@ -14,43 +17,61 @@ public class CustomerService : ICustomerService
         _context = context;
     }
 
-    public List<Customer> GetAll()
+    public Response<List<Customer>> GetAll()
     {
         using var context = _context.Connection();
         string cmd = "SELECT * FROM Customers";
         var customers = context.Query<Customer>(cmd).ToList();
-        return customers;
+        return new Response<List<Customer>>(customers);
     }
 
-    public Customer GetCustomerById(int id)
+    public Response<Customer> GetCustomerById(int id)
     {
         using var context = _context.Connection();
         string cmd = "SELECT * FROM Customers WHERE CustomerId = @CustomerId";
         var customer = context.QueryFirstOrDefault<Customer>(cmd, new { CustomerId = id });
-        return customer;
+        return customer == null
+            ? new Response<Customer>(HttpStatusCode.NotFound, "Customer Not Found")
+            : new Response<Customer>(customer);
     }
 
-    public bool CreateCustomer(Customer customer)
+    public Response<bool> CreateCustomer(Customer customer)
     {
         using var context = _context.Connection();
         var cmd = "INSERT INTO Customers (FullName, Phone, Email) VALUES (@FullName, @Phone, @Email)";
         var response = context.Execute(cmd, customer);
-        return response > 0;
+        return response == 0
+            ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal Server Error")
+            : new Response<bool>(HttpStatusCode.Created, "Customer successfully created");
     }
 
-    public bool UpdateCustomer(Customer customer)
+    public Response<bool> UpdateCustomer(Customer customer)
     {
         using var context = _context.Connection();
+        var existingCustomer = GetCustomerById(customer.CustomerId).Data;
+        if (existingCustomer == null)
+        {
+            return new Response<bool>(HttpStatusCode.NotFound, "Customer not found");
+        }
         var cmd = "UPDATE Customers SET FullName = @FullName, Phone = @Phone, Email = @Email WHERE CustomerId = @CustomerId";
         var response = context.Execute(cmd, customer);
-        return response > 0;
+        return response > 0
+            ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal server error")
+            : new Response<bool>(HttpStatusCode.OK, "Customer successfully updated");
     }
 
-    public bool DeleteCustomer(int id)
+    public Response<bool> DeleteCustomer(int id)
     {
         using var context = _context.Connection();
+        var customer = GetCustomerById(id).Data;
+        if (customer == null)
+        {
+            return new Response<bool>(HttpStatusCode.NotFound, "Customer not found");
+        }
         var cmd = "DELETE FROM Customers WHERE CustomerId = @CustomerId";
         var response = context.Execute(cmd, new { CustomerId = id });
-        return response != 0;
+        return response > 0
+            ? new Response<bool>(HttpStatusCode.InternalServerError, "Internal server error")
+            : new Response<bool>(HttpStatusCode.OK, "Customer successfully deleted");
     }
 }
